Move reused code to objects
