+++ c_src/rocksdb_lib.cpp
@@ -1,4 +1,6 @@
 #include "rocksdb_nif.h"
+#include "rocksdb_nif_resources.h"
+#include "term_prep.h"
 #include "rocksdb/convenience.h"
 #include "table/block_based_table_factory.h"
 
@@ -444,45 +446,19 @@ rocksdb::Status PutTerms(db_obj_resource* rdb,
 			 rocksdb::Slice* value,
 			 std::vector<std::pair<Term, std::vector<Term>>> indices
 			) {
+    rocksdb::Status status;
+    rocksdb::WriteBatch batch;
     // prepare any new indices
     TermPrep tp = TermPrep(indices, key);
     // delete any indicies that may be overwritten by this write
-    TermDelete td = TermDelete(tp.cids_, key);
-    rocksdb::Status status;
-    rocksdb::WriteBatch batch;
-    rocksdb::ReadOptions readoptions;
+    TermDelete td = TermDelete(tp.cids, key);
 
-    //Delete any index history
-    for (auto it = td.index_.begin(); it != td.index_.end(); ++it) {
-	rocksdb::Slice key2term(*it);
-	rocksdb::PinnableSlice value;
-	status = Get(rdb, &readoptions, 1, &key2term, &value);
-	if (status.ok()) {
-	    td.ParseReveseIndex(*it, &value);
-	}
-	batch.Delete(rdb->handles->at(1), key2term);
-    }
-    //Merges for reverse indexes
-    for (auto it = td.rev_index_.begin(); it != td.rev_index_.end(); ++it){
-	batch.Merge(rdb->handles->at(2),
-		    rocksdb::Slice(*it),
-		    rocksdb::Slice(td.posting_));
-    }
+    td.add_to_batch(rdb, batch);
 
     //Put value
     batch.Put(rdb->handles->at(0), *key, *value);
-    //Merge for keeping index history
-    for (auto it = tp.index_.begin(); it != tp.index_.end(); ++it){
-	batch.Put(rdb->handles->at(1),
-		  rocksdb::Slice(it->first),
-		  rocksdb::Slice(it->second));
-    }
-    //Merges for reverse indexes
-    for (auto it = tp.rev_index_.begin(); it != tp.rev_index_.end(); ++it){
-	batch.Merge(rdb->handles->at(2),
-		    rocksdb::Slice(it->first),
-		    rocksdb::Slice(it->second));
-    }
+    tp.add_to_batch(rdb, batch);
+
     //Apply batch according to db type
     if (rdb->type == DB_WITH_TTL) {
 	rocksdb::DBWithTTL* db = static_cast<rocksdb::DBWithTTL*>(rdb->object);
@@ -522,28 +498,15 @@ rocksdb::Status DeleteTerms(db_obj_resource* rdb,
 			    rocksdb::WriteOptions* writeoptions,
 			    rocksdb::Slice* key,
 			    std::vector<Term> cids) {
-    TermDelete td = TermDelete(cids, key);
     rocksdb::Status status;
     rocksdb::WriteBatch batch;
-    rocksdb::ReadOptions readoptions;
+    TermDelete td = TermDelete(cids, key);
+
     //Delete value
     batch.Delete(rdb->handles->at(0), *key);
-    //Delete index history
-    for (auto it = td.index_.begin(); it != td.index_.end(); ++it) {
-	rocksdb::Slice key2term(*it);
-	rocksdb::PinnableSlice value;
-	status = Get(rdb, &readoptions, 1, &key2term, &value);
-	if (status.ok()) {
-	    td.ParseReveseIndex(*it, &value);
-	}
-	batch.Delete(rdb->handles->at(1), key2term);
-    }
-    //Merges for reverse indexes
-    for (auto it = td.rev_index_.begin(); it != td.rev_index_.end(); ++it){
-	batch.Merge(rdb->handles->at(2),
-		    rocksdb::Slice(*it),
-		    rocksdb::Slice(td.posting_));
-    }
+
+    td.add_to_batch(rdb, batch);
+
     //Apply batch according to db type
     if (rdb->type == DB_WITH_TTL) {
 	rocksdb::DBWithTTL* db = static_cast<rocksdb::DBWithTTL*>(rdb->object);
